Added ride generation for history page and moved classes around accordingly
